rename http.rs to admin.rs
