+++ upstairs/src/lib.rs
@@ -36,7 +36,7 @@ use aes_gcm_siv::aead::{AeadInPlace, NewAead};
 use aes_gcm_siv::{Aes256GcmSiv, Key, Nonce, Tag};
 use rand_chacha::ChaCha20Rng;
 
-mod http;
+mod admin;
 mod pseudo_file;
 mod test;
 
@@ -134,7 +134,7 @@ pub struct CrucibleOpts {
     pub cert_pem: Option<String>,
     pub key_pem: Option<String>,
     pub root_cert_pem: Option<String>,
-    pub http: Option<SocketAddr>,
+    pub admin: Option<SocketAddr>,
 }
 
 impl CrucibleOpts {
@@ -2820,7 +2820,7 @@ impl Upstairs {
             cert_pem: None,
             key_pem: None,
             root_cert_pem: None,
-            http: None,
+            admin: None,
         };
         Self::new(
             &opts,
@@ -6456,12 +6456,12 @@ pub async fn up_main(opt: CrucibleOpts, guest: Arc<Guest>) -> Result<()> {
     drop(ds_status_tx);
     drop(ds_reconcile_done_tx);
 
-    // If requested, start the http server on the given address:port
-    if let Some(http) = opt.http {
+    // If requested, start the admin http server on the given address:port
+    if let Some(admin) = opt.admin {
         let upi = Arc::clone(&up);
         tokio::spawn(async move {
-            let r = http::start(&upi, http).await;
-            println!("HTTP task finished with {:?}", r);
+            let r = admin::start(&upi, admin).await;
+            println!("Admin HTTP task finished with {:?}", r);
         });
     }
     /*
